Remove most of damage from unit_generic to its own class
