+++ engine/src/cmd/basecomputer.cpp
@@ -54,6 +54,7 @@ using VSFileSystem::SaveFile;
 #include "gamemenu.h" //network menu.
 #include "audiolib.h"
 #include "vs_math.h"
+#include "damageable.h"
 //#define VS_PI 3.1415926535897931
 
 
@@ -5124,7 +5125,7 @@ void showUnitStats( Unit *playerUnit, string &text, int subunitlevel, int mode, 
     const Unit::UnitJump &uj  = playerUnit->GetJumpStatus();
     const Unit::UnitJump &buj = blankUnit->GetJumpStatus();
     if (!mode) {
-        float maxshield = totalShieldEnergyCapacitance( playerUnit->shield );
+        float maxshield = Damageable::totalShieldEnergyCapacitance( playerUnit->shield );
         if (shields_require_power)
             maxshield = 0;
         PRETTY_ADDU( statcolor+"Recharge: #-c", playerUnit->EnergyRechargeData()*RSconverter, 0, "MJ/s" );
@@ -5629,7 +5630,7 @@ void showUnitStats( Unit *playerUnit, string &text, int subunitlevel, int mode, 
         return;
     if (subunitlevel == 0 && mode == 0) {
         text += "#n##n##c0:1:.5#"+prefix+"[KEY FIGURES]#n##-c";
-        float maxshield = totalShieldEnergyCapacitance( playerUnit->shield );
+        float maxshield = Damageable::totalShieldEnergyCapacitance( playerUnit->shield );
         if (shields_require_power)
             maxshield = 0;
         PRETTY_ADDU( statcolor+"Minimum time to reach full overthrust speed: #-c",
